+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Notable changes to this project.
 
 ## [dev]
 - tests: start testing the cli interface
+- cli: fix `version` on Python2.7
+- added 'liveAuroc' and `validationAuroc` to `get_leaderboard`
 
 ## [1.4.6] - 2019-03-30
 - remove total payments from leaderboard query (#13)